srcs/pyproject.toml
[tool.poetry]
name = "srcs"
version = "0.1.0"
description = ""
authors = ["Eduardo Castro <eduardocastrobritoo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.38.0"
openai = "0.28"
google-generativeai = "^0.8.1"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
sentence-transformers = "^3.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
